+++ src/Endpoint/Endpoints/Promotion.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Onetoweb\Mirakl\Endpoint\Endpoints;
+
+use Onetoweb\Mirakl\Endpoint\AbstractEndpoint;
+
+/**
+ * Promotion Endpoint.
+ */
+class Promotion extends AbstractEndpoint
+{
+    /**
+     * @param array $query = []
+     * 
+     * @return array|NULL
+     */
+    public function list(array $query = []): ?array
+    {
+        return $this->client->get('/api/promotions', $query);
+    }
+}